Refactor project structure, update imports, and improve error handling
